+++ ntex/src/framed/dispatcher.rs
@@ -162,7 +162,7 @@ where
         self
     }
 
-    /// Set connection disconnect timeout in milliseconds.
+    /// Set connection disconnect timeout in seconds.
     ///
     /// Defines a timeout for disconnect connection. If a disconnect procedure does not complete
     /// within this time, the connection get dropped.