+++ ntex/CHANGES.md
@@ -1,5 +1,7 @@
 # Changes
 
+* fix wrong api docs
+
 ## [0.3.12] - 2021-03-18
 
 * http: add per request handler service for http1